requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.16
django-filter==24.3
django-restframework==0.0.1
djangorestframework==3.15.2
importlib_metadata==8.5.0
kombu==5.4.2
Markdown==3.7
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.0
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
zipp==3.21.0
